+++ src/pas/plugins/imio/providers/authentic.py
@@ -10,10 +10,14 @@ Providers which implement the |oauth2|_ protocol.
     Authentic
 """
 from authomatic.providers.oauth2 import OAuth2
+from jwcrypto.jwk import JWKSet
+from jwcrypto.jwt import JWT
+from pas.plugins.imio.utils import authentic_cfg
 from pas.plugins.imio.utils import protocol
 
-import jwt
+import json
 import os
+import requests
 
 
 __all__ = ["Authentic"]
@@ -74,6 +78,10 @@ class Authentic(OAuth2):
         return "{0}/idp/oidc/token/".format(self.base_url)
 
     @property
+    def certs_url(self):
+        return "{0}/idp/oidc/certs/".format(self.base_url)
+
+    @property
     def user_info_url(self):
         return "{0}/idp/oidc/user_info/".format(self.base_url)
 
@@ -89,18 +97,20 @@ class Authentic(OAuth2):
     def _x_user_parser(user, data):
         encoded = data.get("id_token")
         if encoded:
-            payload_data = jwt.decode(
-                encoded,
-                algorithms=["RS256"],
-                options={"verify_signature": False, "verify_aud": False},
-            )
+            __import__("pdb").set_trace()
+            authentic_type = "authentic-agents"
+            hostname = authentic_cfg()[authentic_type]["hostname"]
+            certs_url = "{0}://{1}/idp/oidc/certs/".format(protocol(), hostname)
+            keyset = JWKSet.from_json(requests.get(certs_url).content)
+            jwtcrypto = JWT(jwt=encoded, key=keyset, algs=["RS256"])
+            payload_data = json.loads(jwtcrypto.token.payload)
             if "sub" in payload_data.keys():
                 user.id = payload_data.get("sub")
         if "sub" in data.keys():
             user.username = data.get("preferred_username")
             user.first_name = data.get("given_name")
             user.last_name = data.get("family_name")
-            fullname = u"{0} {1}".format(user.first_name, user.last_name)
+            fullname = "{0} {1}".format(user.first_name, user.last_name)
             if not fullname.strip():
                 user.name = user.username
                 user.fullname = user.username
@@ -112,9 +122,9 @@ class Authentic(OAuth2):
             if len(roles) > 0:
                 app_id = os.environ.get("application_id", "")
                 service_slug = os.environ.get("service_slug", "")
-                if any("{}-admin".format(app_id) in role for role in roles):
+                if any("{0}-admin".format(app_id) in role for role in roles):
                     user.roles.append("Manager")
-                if any("{}-admin".format(service_slug) in role for role in roles):
+                if any("{0}-admin".format(service_slug) in role for role in roles):
                     user.roles.append("Site Manager")
         return user